+++ pkg/providers/property/testdata/TestDataPropertyHostnames/property_hostnames_with_version.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_hostnames" "akaprophosts" {
+  group_id    = "grp_test"
+  contract_id = "ctr_test"
+  property_id = "prp_test"
+  version     = 5
+}