+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## 5.5.0 (xx xx, 2023)
 
+#### FEATURES/ENHANCEMENTS:
+
 * Cloudlets
   * Add `origin_description` field to `akamai_cloudlets_application_load_balancer` resource
+
+* PAPI
+  * Added version support for `akamai_property_hostnames` data source ([I#413](https://github.com/akamai/terraform-provider-akamai/issues/413))
 
 ## 5.4.0 (Oct 31, 2023)